correct errors
Question2: diff_reconstruct = np.sum(np.abs(laplacian.A-L_reconstruct)) is better.

Question 5: the y label of the plot is wrong. It should be "values in u0".

Question 6: The result of Theoretical maximum eigenvalue of combinatorial Laplacians is wrong. The diagonal should be ignored and the argument of axis in function np.sum should be 1.